% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_size.R
\name{benchmark_size}
\alias{benchmark_size}
\alias{benchmark_size.default}
\alias{benchmark_size.data.frame}
\title{Benchmark Expressions Against Variably Sized Data}
\usage{
benchmark_size(object, ..., size = c(10, 100, 1000), times = 100,
  unit = "ms")

\method{benchmark_size}{default}(object, ..., size = c(10, 100, 1000),
  times = 100, unit = "ms")

\method{benchmark_size}{data.frame}(object, ..., size = c(10, 100, 1000),
  times = 100, unit = "ms")
}
\arguments{
\item{...}{expressions to benchmark}

\item{size}{\code{numeric}, sample sizes to iterate over.}

\item{times}{\code{numeric}, how many times to run the code in
\code{microbenchmark}}

\item{unit}{\code{character}, specifies the unit of time to measure.}

\item{x}{An atomic vector or a data frame}
}
\description{
Determining the optimal expression to use is not always
obvious, and the size of the data frame may make a difference in what
is the fastest solution. This function manages the benchmarking over
datasets of different sizes and returns the results to facilitate
a decision about which solution to use.
}
