% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_data.R
\name{bootstrap_data}
\alias{bootstrap_data}
\title{Lists of Bootstrapped Data Frames}
\usage{
bootstrap_data(data, groups = character(0), B = 40, seed = numeric(0),
  cl = NULL)
}
\arguments{
\item{data}{The data frame to be bootstrapped}

\item{groups}{A character vector giving the names of the columns in the data
frame that define groups or clusters that should be sampled together.
Use \code{character(0)} or \code{NULL} for each row as a distinct entity.}

\item{B}{\code{numeric(1)}, an integer-ish value giving the number of
data frames to return.}

\item{seed}{A numeric value with which the random number generator may 
be seeded.  Using \code{character{0}} or \code{NULL} causes the 
function to behave off of whatever exists in the environment at present.}
}
\description{
Bootstrapping data frames is a critical part of developing 
  bootstrapped estimates from models.  This function provides a basic 
  interface to generate the bootstrapped data to feed to a modeling function.
}
\section{Functional Requirements}{

\enumerate{
  \item Accepts objects that inherits the \code{data.frame} class.
  \item Objects returned in the list have the same class as what was input.
  \item Returns a list of length \code{B}
  \item cl An object that inherits the \code{cluster} class. Used for
    parallel computation. 
}
}
\author{
Benjamin Nutter
}

