% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constants.R
\name{ControlChartConstants}
\alias{A1}
\alias{A2}
\alias{A3}
\alias{A4}
\alias{A6}
\alias{A9}
\alias{B3}
\alias{B4}
\alias{ControlChartConstants}
\alias{D3}
\alias{D4}
\alias{D5}
\alias{D6}
\alias{c2}
\alias{c4}
\alias{d2}
\alias{d3}
\alias{d4}
\title{Control Chart Constants}
\source{
Donald J. Wheeler and David S. Chambers, \emph{Understanding Statistical Process Control, Second Edition},
SPC Press, Knoxville, TN, 1992. ISBN: 0-945320-13-2.\cr
For c2 and c4, see page 393. \cr
For A2, D3, and D4, see page 394. \cr
For A4, D5, and D6, see page 395. (d4 not yet defined) \cr
For A1, B3, and B4, see page 396. \cr
For A3, B3, and B4, see page 397. \cr
For A6 and A9, see pages 398-399.


Emanuel P. Barbosa, Flavio M. M. Barros, Elias de Jesus Goncalves and Daniela R. Recchia (2014).
\emph{IQCC: Improved Quality Control Charts}. R package version 0.6. http://CRAN.R-project.org/package=IQCC
d2 and d3 are copied from their package.
}
\usage{
A1(n)

A2(n)

A3(n)

A4(n)

A6(n)

A9(n)

B3(n, rms = TRUE)

B4(n, rms = TRUE)

c2(n)

c4(n)

d2(n)

d3(n)

d4(n)

D3(n)

D4(n)

D5(n)

D6(n)
}
\arguments{
\item{n}{Subgroup sample size}

\item{rms}{Determines if the root mean square deviation is used or the standard deviation.
  The root mean deviation divides by \code{n} while the standard deviation divides by
  \code{n-1}.}
}
\description{
Functions to calculate or reference control chart constants.
}
\details{
All of the constants are directly calculated within R except for d4, A6, and A9.  I
  have been unable to determine a way to calculate the d4 constant from the \code{ptukey}
  function, nor have I figured out the relationship between the skewness of the studentized
  range distribution and d4.  I have not found formulae for A6 and A9.  Instead, the values of
  these constants are copied out of Wheeler and Chambers and referenced out of the a data
  frame.  In cases where the supplied value of \code{n} is not found in the reference table,
  the constant for the largest tabulated n that is smaller than \code{n} is used.
}
\author{
Benjamin Nutter (except for d2 and d3)

Barbosa et. al (d2, d3) (vectorized by Nutter)
}

