% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/countConsecutive.R
\name{countConsecutive}
\alias{countConsecutive}
\title{Count Runs of Observations Satisfying a Condition}
\usage{
countConsecutive(x, condition, maxIndex = TRUE, ...)
}
\arguments{
\item{x}{A numeric vector of observations.}

\item{condition}{An unevaluated logical expression based on \code{x}.}

\item{maxIndex}{Logical. When \code{TRUE}, each element is the maximum run
length of each unique run.  When \code{FALSE}, each element is the element's
individual run count.}

\item{...}{Additional arguments to be passed to \code{plyr::ddply}
  (especially for running in parallel).}
}
\value{
A vector the length of \code{x} giving each elements run length or
  run count.
}
\description{
Shewhart's rules provide conditions under which a process can
  be considered to be out of control. \code{countConsecutive} generates
  an index by which runs can be systematically identified and can
  facilitate identifying these observations on control charts.
}
\details{
Shewhart's rules for identifying out of control processes are:
\enumerate{
  \item Any one point is more than three standard units from the central line
  \item Any two consecutive points are more than two standard units in the
        same direction from the central line
  \item A run of ___ consecutive points are more than one standard units in
        the same direction from the central line
  \item A run of eight consecutive points are on the same side of the
        central line.
}
}
\author{
Benjamin Nutter
}

