% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_contents.R
\name{data_contents}
\alias{data_contents}
\title{Summarize Data}
\usage{
data_contents(data)
}
\arguments{
\item{data}{A data frame to be summarized}
}
\value{
The columns of the overview table include:
\itemize{
  \item Variable The variable name
  \item Label The variable label
  \item Type The variable class
  \item Missing The number of missing values
  \item Prop.Miss The proportion of missing values.
}

The columns of the numeric summary table include:
\itemize{
  \item Factor The variable label
  \item N The number of non-missing values
  \item Missing The number of missing values
  \item Mean The variable mean
  \item SD The variable Standard Deviation
  \item LCL The lower 95% confidence limit for the mean
  \item UCL The upper 95% confidence limit for the mean
  \item Min The variable minimum
  \item P25 The variable 25th percentile
  \item MEDIAN The variable median
  \item P75 The variable 75th percentile
  \item MAX the variable maximum
  \item CV The variable coefficient of variation
}

The columns of the categorical summary table include:
\itemize{
  \item Variable Variable label
  \item Level The level of the variable
  \item Missing The number of missing values
  \item Freq The frequency of the level
  \item Rel.Freq The relative frequency of the level
  \item Cum.Freq The cumulative frequency for the variable
  \item Rel.Cum.Freq The cumulative relative frequency for the table.
}  

The columns of the date summary table include:
\itemize{
  \item Variable The variable label
  \item Missing The number of missing values
  \item Min The minimum date
  \item Max The maximum date
} 
  
The columns of the other variables summary table include:
\itemize{
  \item Label The variable label
  \item Missing The number of missing values
  \item Type The variable class
}
}
\description{
Prepare a set of tables to describe a data set.  Five tables 
  are available giving an overview of the variables and summaries of numeric,
  categorical, date, and other variables (usually character).
}
\details{
The overview summary table is an approximation of the SAS PROC 
CONTENTS output.  

The numeric summary is generated using the \code{univ} function from the
\code{lazyWeave} package.
}
\author{
Benjamin Nutter
}
