% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataContents.R
\name{dataContents}
\alias{dataContents}
\title{Summarize Data}
\usage{
dataContents(data, round = 2, report = TRUE, print_tabs = c("overview",
  "numeric", "categorical", "date", "other"), overview_args = list(align =
  c("left", "left", "left", "center", "center"), caption = "Variable Overview"),
  numeric_args = list(align = c("left", rep("center", 12)), caption =
  "Summary of Numerical Variables"), categorical_args = list(align = c("left",
  "left", rep("center", 5)), caption = "Summary of Categorical Variables"),
  date_args = list(align = c("left", rep("center", 3)), caption =
  "Summary of Date Variables"), other_args = list(align = c("left", "center",
  "center"), caption = "Summary of Other Variables"), ...)
}
\arguments{
\item{data}{A data frame to be summarized}

\item{round}{The number of significant digits to which numerical values
will be rounded in the summaries.}

\item{report}{Logical.  Indicates if the results are printed to a report
(using \code{lazyWeave}) or returned to the environment.}

\item{print_tabs}{A character vector giving the tables to be returned.
Partial matches are accepted.}

\item{overview_args}{list of additional arguments to be passed to
\code{lazy.matrix} for the overview table.}

\item{numeric_args}{list of additional arguments to be passed to
\code{lazy.matrix} for the numeric summary table.}

\item{categorical_args}{list of additional arguments to be passed to
\code{lazy.matrix} for the categorical summary table.}

\item{date_args}{list of additional arguments to be passed to
\code{lazy.matrix} for the date summary table.}

\item{other_args}{list of additional arguments to be passed to
\code{lazy.matrix} for the other variables summary table.}

\item{...}{Arguments to be passed to other methods. Currently ignored.}
}
\value{
The columns of the overview table include:
\enumerate{
  \item Variable The variable name
  \item Label The variable label
  \item Type The variable class
  \item Missing The number of missing values
  \item Prop.Miss The proportion of missing values.
}

The columns of the numeric summary table include:
\enumerate{
  \item Factor The variable label
  \item N The number of non-missing values
  \item Missing The number of missing values
  \item Mean The variable mean
  \item SD The variable Standard Deviation
  \item LCL The lower 95% confidence limit for the mean
  \item UCL The upper 95% confidence limit for the mean
  \item Min The variable minimum
  \item P25 The variable 25th percentile
  \item MEDIAN The variable median
  \item P75 The variable 75th percentile
  \item MAX the variable maximum
  \item CV The variable coefficient of variation
}

The columns of the categorical summary table include:
\enumerate{
  \item Variable Variable label
  \item Level The level of the variable
  \item Missing The number of missing values
  \item Freq The frequency of the level
  \item Rel.Freq The relative frequency of the level
  \item Cum.Freq The cumulative frequency for the variable
  \item Rel.Cum.Freq The cumulative relative frequency for the table.
}

The columns of the date summary table include:
\enumerate{
  \item Variable The variable label
  \item Missing The number of missing values
  \item Min The minimum date
  \item Max The maximum date
}

The columns of the other variables summary table include:
\enumerate{
  \item Label The variable label
  \item Missing The number of missing values
  \item Type The variable class
}
}
\description{
Prepare a set of tables to describe a data set.  Five tables
  are available giving an overview of the variables and summaries of numeric,
  categorical, date, and other variables (usually character).
}
\details{
The overview summary table is an approximation of the SAS PROC
CONTENTS output.

The numeric summary is generated using the \code{univ} function from the
\code{lazyWeave} package.
}
\author{
Benjamin Nutter
}

