% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_variable.R
\name{dummy_variable}
\alias{dummy_variable}
\title{Create Columns of Dummy Variables}
\usage{
dummy_variable(data, x, lev = NULL, var_name = FALSE, as_logical = FALSE)
}
\arguments{
\item{data}{An object that inherits data frame.}

\item{x}{A \code{character(1)} naming a variable in \code{data}}

\item{lev}{A vector of length 1. The value in \code{x} to use as the 
reference value.}

\item{var_name}{\code{logical(1)}. When \code{TRUE}, the value of 
\code{x} is prepended to the level, separated by an underscore.}

\item{as_logical}{\code{logical(1)}. When \code{TRUE}, the dummy variables
are returned as logical values.  By default, they are returned as 0/1 values.}
}
\description{
Create a data frame of dummy variables.  There will be 
  n-1 variables in the data frame, where n is the number of levels in 
  the factor.
}
\section{Functional Requirements}{


In the requirements, \code{x} is used as shorthand for \code{data[["x"]]}

\enumerate{
  \item Accepts an object that inherits \code{data.frame}
  \item Accepts a \code{character(1)} naming a variable in \code{data}
  \item Casts an error when \code{x} is not in \code{data}
  \item Accepts a length 1 vector that names a value in \code{x} to use as 
    the reference value.  
  \item Casts and error if \code{lev} is not a value in \code{x}
  \item Provide an option to prepend the variable name to the front of 
    the level name.
  \item Provide an option to return dummy values as logical values.
}
}
\examples{
# Dummy Variables as 0/1
dummy_variable(iris, "Species")

# Dummy Variables as 0/1 with prefix
dummy_variable(iris, "Species", var_name = TRUE)

# Dummy Variables as TRUE/FALSE
dummy_variable(iris, "Species", as_logical = TRUE)

# Dummy Variables as TRUE/FALSE with prefix
dummy_variable(iris, "Species", var_name = TRUE, as_logical = TRUE)

# Dummy Variables as 0/1 with alternate reference
dummy_variable(iris, "Species", lev = "virginica")

}

