% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_reference.R
\name{factor_reference}
\alias{factor_reference}
\title{Generate a list of reference tables for factors in a data set}
\usage{
factor_reference(data)
}
\arguments{
\item{data}{A data frame-like object.}
}
\description{
Creates a list consisting of one data frame for each factor in a 
  data set.  Each table has two columns; \code{level} and \code{label}.
}
\section{Functional Requirements}{

\enumerate{
  \item Accept an object that inherits the \code{data.frame} class
  \item Return a list of data frames, one for each factor in \code{data}
  \item Return a message indicating when no factors are found.
}
}

