% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/firstRecord.R
\name{FirstLastRecord}
\alias{FirstLastRecord}
\alias{firstRecord}
\alias{lastRecord}
\alias{nthRecord}
\title{First, Last, and nth Records}
\usage{
firstRecord(data, id, ...)

lastRecord(data, id, ...)

nthRecord(data, id, ..., n = 1)
}
\arguments{
\item{data}{A data frame to be reduced to the first row per identifiers}

\item{id}{a vector of identifiers}

\item{...}{additional variables by which the data should be sorted prior
to selecting the first per identifier set.}

\item{n}{The record number to pull from each ID.}
}
\description{
Extracts specific records from within an id in a repeated
  measures data set.
}
\details{
The data frame will first be sorted by the variables in \code{id}
  and then by the variables in \code{...}.

  Sorting is performed by the \code{dplyr::arrange} function, and accepts
  use of the \code{desc} to sort a variable in descending order.  In both
  \code{firstRecord} and \code{lastRecord}, sorting should be done in the
  natural sort order of the variable; the last record is identified using
  the \code{duplicated} function with \code{fromLast=TRUE}.

  For \code{nthRecord}, using \code{desc} around each of the variables in
  \code{...} will provide the nth-from-last record for each id.
}
\author{
Benjamin Nutter
}
\seealso{
\code{order}, \code{duplicated}
}

