% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gest_age.R
\name{gest_age}
\alias{gest_age}
\title{Calculate Gestational Age}
\source{
\url{http://www.acog.org/-/media/Departments/Patient-Safety-and-Quality-Improvement/201213IssuesandRationale-GestationalAgeTerm.pdf }
}
\usage{
gest_age(edc, adc, result = c("both_string", "both_df", "week", "day",
  "week_decimal", "day_decimal"))
}
\arguments{
\item{edc}{The estimated date of confinement (sometimes called the estimated
date of delivery).  Must be an object suitable 
for the \code{difftime} function (such as \code{Date}, \code{POSIXct},
etc.) or the calculation will fail.}

\item{adc}{The date for which the gestational age needs to be calcualted.
This may be the current date, the date of a prenatal care visit, or the
actual date of delivery.  As with \code{edc}, this must be an object
suitable for \code{difftime}.}

\item{result}{A character option that determines the nature of the result
returned.  See Details.}
}
\description{
Calculate the gestational age of a child or fetus based on the
  estimated date of confinement (or estimated date of delivery).
}
\details{
Gestational age is commonly reported as [weeks]/[days] where 
  [weeks] is an integer between 0 and 42 (though in some cases is may be 
  larger), and days is an integer between 0 and 6.  Thus, 39/2 is 39 weeks 
  and 2 days into pregnancy.  The commonly accepted standard length of 
  pregnancy is 40 weeks.
  
  The \code{result} argument controls how the output is reported to the 
  user with the following options.
  \enumerate{
    \item{both_string }{A character string in the format "[week]/[day]"}
    \item{both_df }{A data frame with a column for week and a column for 
                    day}
    \item{week }{A numeric value of the weeks gestational age}
    \item{day }{A numeric value of the day component of the gestational
                age (an integer between 0 and 6)}
    \item{week_decimal }{A numeric value of the total weeks gestational
                         age (\code{week + (day / 7)})}
    \item{day_decimal }{A numeric value of the total days gestational 
                        age (\code{week * 7 + day})}
  }
}
\section{Functional Requirements}{

\enumerate{
  \item Accept two date-like objects to calculate gestational age.
  \item Return the number of weeks, the number of days within the week,
        both, the number of decimal weeks, or the number of decimal days.
}
}

\examples{
edc <- as.Date(c("6/18/2010", "10/22/2010"), format="\%m/\%d/\%Y")
adc <- as.Date(c("5/22/2010", "10/18/2010"), format="\%m/\%d/\%Y")
gest_age(edc, adc, "both_string")
gest_age(edc, adc, "week")
gest_age(edc, adc, "day")

}
\author{
Benjamin Nutter
}
