% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gestAge.R
\name{gestAge}
\alias{gestAge}
\title{Calculate Gestational Age}
\source{
\url{http://www.acog.org/-/media/Departments/Patient-Safety-and-Quality-Improvement/201213IssuesandRationale-GestationalAgeTerm.pdf }
}
\usage{
gestAge(edc, adc, result = c("both", "week", "day"))
}
\arguments{
\item{edc}{The estimated date of confinement (sometimes called the estimated
date of delivery).  Must be an object suitable 
for the \code{difftime} function (such as \code{Date}, \code{POSIXct},
etc.) or the calculation will fail.}

\item{adc}{The date for which the gestational age needs to be calcualted.
This may be the current date, the date of a prenatal care visit, or the
actual date of delivery.  As with \code{edc}, this must be an object
suitable for \code{difftime}.}

\item{result}{A character option that determines the nature of the result
returned.  \code{"both"} returns a character string in the format
\code{"[weeks]/[days]"}; \code{"week"} returns the integer value of the 
weeks component; \code{"day"} returns the integer value of the days 
component.}
}
\description{
Calculate the gestational age of a child or fetus based on the
  estimated date of confinement (or estimated date of delivery).
}
\details{
Gestational age is commonly reported as [weeks]/[days] where 
  [weeks] is an integer between 0 and 42 (though in some cases is may be 
  larger), and days is an integer between 0 and 6.  Thus, 39/2 is 39 weeks 
  and 2 days into pregnancy.  The commonly accepted standard length of 
  pregnancy is 40 weeks.
}
\examples{
edc <- as.Date(c("6/18/2010", "10/22/2010"), format="\%m/\%d/\%Y")
adc <- as.Date(c("5/22/2010", "10/18/2010"), format="\%m/\%d/\%Y")
gestAge(edc, adc, "both")
gestAge(edc, adc, "week")
gestAge(edc, adc, "day")

}
\author{
Benjamin Nutter
}
