% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_km.R
\name{gg_km}
\alias{gg_km}
\alias{gg_km.survfit}
\title{Kaplan-Meier Survival Plots using \code{ggplot}}
\usage{
gg_km(object, times = NULL, ...)

\method{gg_km}{survfit}(object, times = NULL, conf_type = c("none",
  "region", "bar", "band"), cum_inc = FALSE, plot_censor = FALSE,
  n_risk = FALSE, n_event = FALSE, region_alpha = 0.5, bar_offset = 1,
  band_linetype = 2, censor_shape = 4, gg_expr = list(), ...)
}
\arguments{
\item{object}{An object of class \code{survfit}, \code{coxph}, or \code{cph}. 
\code{coxph} and \code{cph} objects are coerced to \code{survfit} objects.}

\item{times}{A numeric vector of time at which ticks on the x-axis should be drawn.
These times are also used for determining where to draw confidence bars, 
number of subjects and risk, and number of cumulative events, when any of 
these are chosen.}

\item{...}{Additional arguments to pass to other methods. Currently ignored.}

\item{conf_type}{A character value denoting the method to use for displaying confidence
limts.  See Details.}

\item{cum_inc}{\code{logical(1)} indicating if the plot should be drawn as a 
cumulative incidence plot.  Cumulative incidence is calcualated as 1 - survival.}

\item{plot_censor}{\code{logical(1)} indicating if times at which censored observations 
are recorded should be plotted.}

\item{n_risk}{\code{logical(1)} indicating if the number at risk should be displayed
below the plot.}

\item{n_event}{\code{logical(1)} indicating if the cumulative number of events should 
be displayed below the plot.}

\item{region_alpha}{\code{numeric(1)} giving the alpha (transparency) value for confidence
regions.}

\item{bar_offset}{\code{numeric(1)}. Specifies the units by which confidence bars are 
separated along the x-axis. Defaults to \code{1}.}

\item{band_linetype}{May be either an integer or character of length 1 specifying the line
type to be used in confidence bands.}

\item{censor_shape}{May be either an integer or character of length 1 specifying the 
plotting character to be used for censored observations.}

\item{gg_expr}{A list of layers to add to the primary plot.  See Details.}
}
\description{
Kaplan-Meier plot for survival in a time-to-event analysis with
  multiple display options.
}
\details{
Confidence limits may be omitted (the default) or drawn as regions (shaded regions extending 
    from the lower to the upper confidence limits); bars (vertical bars extending from 
    the lower to the upper confidence limits at \code{times}; or bands (step function lines
    drawn through the lower and the upper confidence limits, as in \code{stats::plot.survfit}).
    
When \code{n_risk} or \code{n_event} is \code{TRUE}, a plot is returned through 
\code{gridExtra::gridArrange}.  When this occurs, the plot object is not directly 
modifiable.  Any additional layers that need to be added to the graph must be passed as 
a list to \code{gg_expr}.
}
\section{Functional Requirements}{

\enumerate{
  \item Accept \code{survfit}, \code{coxph}, and \code{cph} objects.
  \item Allow user to specify the times displayed on the x-axis.  If no times are
        selected, use the default times from the \code{survfit} object.
  \item Allow the user to select no confidence limits, confidence regions, confidence
        bars, or confidence bands.
  \item Display confidence bars only at the times requested by the user
  \item Allow user control over the offset of the confidence bars
  \item Include an option to display as a cumulative incidence plot
  \item Include an option to display number of subjects at risk. Only display for 
        times requested by the user.
  \item Include an option to display cumulative events. Only display for 
        times requested by the user.
  \item Provide a list argument that may be used to modify the content of the plot
        when a simple \code{ggplot} object cannot be returned.
}
}

\author{
Benjamin Nutter
}
