% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_km_data.R
\name{gg_km_data}
\alias{gg_km_data}
\alias{gg_km_data.survfit}
\alias{gg_km_data.summary.survfit}
\title{Construct Data for Use in Kaplan-Meier Plots}
\usage{
gg_km_data(object, ...)

\method{gg_km_data}{survfit}(object, ...)

\method{gg_km_data}{summary.survfit}(object, offset = NULL, ...)
}
\arguments{
\item{object}{Either a \code{survfit} or \code{summary.survfit} object. The 
\code{survfit} object is used to build the base plot.  The 
\code{summary.survfit} object is used for formatting the x-axis, N at risk,
N events, and confidence bars.}

\item{...}{Additional arguments to pass to other methods.}

\item{offset}{Offset values to be used for confidence bars.}
}
\description{
Generates data sets suitable for use in \code{ggplot} style 
  Kaplan-Meier plots
}
\details{
The \code{survfit} method forces the data in each group to start at time 0 (unless
  a negative time exists)
  
The \code{summary.survfit} method only returns times requested by the user (or the default 
  times from the \code{survfit} object).
}
\section{Functional Requirements}{

\enumerate{
  \item Accept either \code{survfit} or \code{summary.survfit} objects
  \item Return a tidy dataframe suitable for use in \code{ggplot} style grpahics
  \item Force the times in \code{survfit} objects to begin at time 0, unless a negative
        time is found in the object as received
  \item Remove the \code{"x="} prefix from the \code{strata} column.
}
}

\author{
Benjamin Nutter
}
