% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSurvGraph.R
\name{ggSurvGraph}
\alias{ggSurvGraph}
\title{Kaplan-Meier Survival Graphs}
\usage{
ggSurvGraph(object, times, cum.inc = FALSE, conf.bar = TRUE,
  offset.scale = 1, n.risk = FALSE, n.event = FALSE, gg_expr)
}
\arguments{
\item{object}{A fitted survival object, such as a survfit model or a data 
frame containing the information for the graph. If a data frame, the 
columns must be named time, surv, lower, upper, n.risk, n.event, group. 
(see details for more information).}

\item{times}{Vector of times at which confidence bars should be drawn. 
Defaults to all event times.}

\item{cum.inc}{If \code{TRUE}, a cumulative incidence plot is drawn instead 
of a survival plot.}

\item{conf.bar}{If \code{TRUE}, vertical bars denoting the confidence 
limits are drawn.}

\item{offset.scale}{the scale of time units to offset the bars from the 
specified interval. For example, if a bar should be placed at 25 months, 
and scale=.5, the bars will be placed at 24.5 and 25.5 months. If scale=1, 
the bars will be placed at 24 months and 26 months. Additional groups are 
placed further out on the same scale}

\item{n.risk}{A logical indicating whether the number of subjects at 
risk should be printed at the bottom of the graph. Defaults to \code{FALSE} 
for no printing.}

\item{n.event}{A logical indicating whether the cumulative number of events 
should be printed at the bottom of the graph. Defaults to \code{FALSE} 
for no printing.}

\item{gg_expr}{A list of expressions to be added to the initial \code{ggplot}
command.}
}
\description{
Creates a graph of one or more survival curves with vertical 
  bars to represent confidence intervals. Uses the \code{ggplot} system 
  instead of base graphics.
}
\details{
The function plots the Kaplan-Meier survival curves of the provided 
  object or data. It then places vertical bars the length of the confidence 
  interval at that point at the distances specified by span.
  
  \code{ggSurvGraph} operates my creating a data frame of values to be 
  plotted, even if a survfit object is passed to the function. Thus, 
  \code{SurvGraph} does not utilize the options in \code{plot.survfit.} 
  Those accustomed to plotting Kaplan-Meier curves with \code{plot.survfit} 
  will not necessarily get the same results with \code{ggSurvGraph}.
  
  When a data frame is passed to \code{ggSurvGraph} the columns of data must 
  have the named elements: time, surv, lower, upper, n.risk, n.event 
  [,strata], where

  When a data frame is passed to SurvGraph the columns of data must have the named elements:
  \code{time, surv, lower, upper, n.risk, n.event [,strata]}, where 
  \tabular{ll}{
    \code{time}  \tab denotes the survival time\cr
    \code{surv}  \tab denotes the probability of survival at \code{time}\cr
    \code{lower} \tab denotes the lower confidence limit for \code{surv}\cr
    \code{upper} \tab denotes the upper confidence limit for \code{surv}\cr
    \code{n.risk}\tab denotes the number at risk at \code{time}\cr
    \code{n.event}\tab denotes the number of events occuring between \code{time} 
      and the previous \code{time}.\cr
    \code{strata}\tab is an optional variable that allows for simultaneous plotting of stratified
      survival curves.  Each observation in the data set should have a group 
      associated with it if \code{group} is supplied.\cr
  }
  
  \code{SurvGraph} is not capable of handling cox models.
}
\section{Internal Data}{

Additional layers may be added to the plot using the \code{gg_expr} 
argument.  The layers should use the data produced internal to the function.
There are two data sets created.  \code{survRaw} is generated from the 
output of \code{summary(object)} without a \code{times} argument.  The
Kaplan-Meier curve is based on this data frame as it contains the information
for the full curve.  A second data frame, called \code{survData}, is generated
from the output of \code{summary(object, times=[times])} and is used to 
draw the confidence bars.  

Contents of \code{survRaw}
\itemize{
  \item{time} The time elapsed since the index time
  \item{n.risk} The number at risk at time
  \item{n.event} The number of events between time[t-1] and time[t].
  \item{n.censor} The number of censored observations between time[t-1] and time[t]
  \item{surv} The survival proportion at time[t]
  \item{lower} The lower confidence limit of survival at time[t]
  \item{upper} The upper confidence limit of survival at time[t]
  \item{strata} Group assignment
  \item{cum.evt} The cumulative events at time[t]
  \item{next.time} The next observed time in the object.  This is useful for 
    plotting confidence bands.
}

The contents \code{survData} are identical, except it lacks the \code{next.time}
variable.
}

\examples{
\dontrun{ 
  library(survival)
  fit <- survfit(Surv(time, status) ~ x, data=aml)
  ggSurvGraph(fit)
  ggSurvGraph(fit, offset.scale=2, n.risk=TRUE) 
  
  #* changing the linetype:
  #* Because the graph is based on the output from the survfit object,
  #* the aesthetics passed to the plot must be based on the variable names
  #* from that object, not the original data frame.  The linetype and 
  #* colour will be based on the 'strata'.
  ggSurvGraph(fit, n.risk=TRUE,
              gg_expr=list(aes(linetype=strata)))
              
  #* Confidence Bands can be obtained using geom_rect()
  ggSurvGraph(fit, conf.bar=FALSE, times=seq(0, 60, by=12),
              gg_expr=list(geom_rect(aes(xmin=time, xmax=next.time,
                                         ymin=lower, ymax=upper,
                                         fill=strata),
                                     alpha=.25, linetype=0)))
}

}
\author{
Benjamin Nutter
}
