% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_levels.R
\name{group_levels}
\alias{group_levels}
\title{Consolidate Levels in a Factor}
\source{
A Stack Overflow question prompted me to write this function:
\url{http://stackoverflow.com/questions/42423022/function-to-determine-proper-group/42423605}
}
\usage{
group_levels(x, ...)
}
\arguments{
\item{x}{A vector.}

\item{...}{Additional named arguments.  The name should reflect the desired group 
name, and should be accompanied by a character vector naming the values to 
map to that group. A name associated with \code{NULL} will map any values 
not declared in other arguments.  See Examples.}
}
\value{
A factor variable with the same length of \code{x}.
}
\description{
Simplifies the consolidation of multiple factor levels into 
  a smaller number of levels.
}
\section{Functional Requirements}{

\enumerate{
  \item Return an object of class \code{factor} with the same length as \code{x}.
  \item All arguments passed to \code{...} must be vectors or \code{NULL}.
  \item No more than one argument passed to \code{...} may be \code{NULL}.
}
}
\examples{
# From the Stack Overflow Question that inspired this function:
DF <- data.frame(Members = c("Eva",  "Charlie1", "Fred",     "Charlie2", 
                             "Adam", "Eva",      "Charlie2", "David", 
                             "Adam", "David",    "Charlie1"))
                  
# Using `Group3 = NULL` means that anything that isn't defined in 
# `Group1` or `Group2` will be assigned to `Group3`

group_levels(DF$Members, 
             Group1 = c("Eva", "Adam"), 
             Group2 = c("Charlie1", "Charlie2"),
             Group3 = NULL)

# Excluding `Group3 = NULL` means that anything that isn't defined in
# `Group1` or `Group2` will be assigned `NA`

group_levels(DF$Members, 
             Group1 = c("Eva", "Adam"), 
             Group2 = c("Charlie1", "Charlie2"))
}
\author{
Benjamin Nutter
}
\seealso{
\code{\link{levels}}
}

