% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_names_to_column.R
\name{list_names_to_column}
\alias{list_names_to_column}
\title{Make list names a column name}
\usage{
list_names_to_column(list_data, new_col)
}
\arguments{
\item{list_data}{a named list of data frames.}

\item{new_col}{\code{character(1)}, the name of the new column in the data 
frame.}
}
\value{
A list of data frames. Each data frame has an additional column given by 
\code{new_col}.
}
\description{
Operations on a list of data frames, such as reading in
  multiple files, do not always place an identifier for the list in the
  data frame itself.  The function provided here takes the name and places
  it in the data frame as a column. This may be accomplished with an 
  \code{mapply} call, but who wants to write those eight lines of code?
}
\section{Functional Requirements}{

\enumerate{
 \item Return a list of data frames.
 \item Each data frame in the list has a new column named by \code{new_col}
 \item Cast an error if \code{list_data} is not a named list of data frames.
 \item Cast an error if \code{new_col} is not a \code{character(1)}
 \item Cast an error if \code{new_col} exists in any element of 
   \code{list_data}.
}
}

