% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_any.R
\name{lrt_any}
\alias{lrt_any}
\title{Likelihood Ratio Tests for Models of any Class}
\usage{
lrt_any(fit0, fit1)
}
\arguments{
\item{fit0}{A fit object with a \code{LogLik} method.}

\item{fit1}{A fit object with a \code{LogLik} method.}
}
\description{
A form of the likelihood ratio test that does not assume 
  the models are of the same form.  The motivation is to compare fixed
  effects models to mixed effects models to determine if the random 
  effects improve the fit.  Note, in most cases, it probably isn't wise
  to perform likelihood ratio tests between models of different types.
  
  The other advantage to this function is it always performs a likelihood
  ratio test. The most common way to perform this test is to use the 
  \code{anova} function. However, since \code{anova} is a generic, there 
  is not guarantee that what is returned is a likelihood ratio test.
}
\examples{
fit0 <- lm(mpg ~ qsec, data = mtcars)
fit1 <- lm(mpg ~ qsec + wt, data = mtcars)
fit2 <- lm(mpg ~ qsec + wt + disp, data = mtcars)

lrt_any(fit0, fit1)
lrt_any(fit1, fit2)

}
