% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_regex.R
\name{month_regex}
\alias{month_regex}
\title{Regular Expression for Extracting Month Names from a String}
\usage{
month_regex(abbreviated = FALSE)
}
\arguments{
\item{abbreviated}{\code{logical(1)}. Should the abbreviated month names be 
used?}
}
\description{
Generates the regular expression string that will
  match any month name in a character string.  This can 
  be helpful for identifying month names in non standard date
  patterns.
}
\section{Functional Requirements}{

\enumerate{
  \item Return a character vector that will match the month names.
  \item Permit the user to request abbreviated month names.
}
}

