% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_to_word.R
\name{number_to_text}
\alias{number_to_text}
\alias{translate_hundred}
\title{Convert Numbers to Text/Word Strings}
\usage{
number_to_text(x)

translate_hundred(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Reprints numeric representations as printed words.  For example, '10' becomes
  'ten' and '28' becomes 'twenty eight'.
}
\details{
This will fail whenever R forces the number to be represented in 
  scientific notation. There is some relief from this as 
  \code{number_to_word} will attempt to reset the \code{scipen} option to 
  prevent the use of scientific notation, but sometimes R insists on
  doing things its own way.  I'm sure there's a good, succinct explanation 
  for when to expect this to fail, but since it is mostly likely to happen 
  at preposterously large values (into the trillions), I'm not too concerned 
  about because let's face it--at that point, you should probably reconsider 
  if you really want to print out the number as text.  Do note, however,
  that exact powers of 10 that exceed one billion are problematic.

  \code{number_to_word} also starts to lose precision in the argument check at 
  4,503,599,627,370,496 (four quadrillion). Precision in the translation 
  appears to be lost for values larger than 9,999,999,999,998. (almost ten
  quadrillion).
}
\section{Functional Requirements}{

\enumerate{
  \item return a character vector the same length as \code{x}
  \item Cast an error if \code{x} is not integerish
  \item Cast a warning if any value in \code{x} is larger than 
    9,999,999,999,998
}
}

\examples{
number_to_text(0:20)

# Largest value before loss of precision in modulo
number_to_text(4503599627370496)
number_to_text(4503599627370497)

# Limits of precision on the translation
number_to_text(9999999999999998)
number_to_text(9999999999999999)  # Incorrectly translated
number_to_text(10000000000000000) # Correctly translated
number_to_text(10000000000000001) # Incorrectly translated
}
\seealso{
\code{\link{word_to_number}}
}
\author{
Benjamin Nutter
}
