% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numberToWord.R
\name{numberToWord}
\alias{numberToWord}
\alias{translateHundred}
\title{Convert Numbers to Text/Word Strings}
\usage{
numberToWord(x)

translateHundred(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Reprints numeric representations as printed words.  For example, '10' becomes
  'ten' and '28' becomes 'twenty eight'.
}
\details{
This will fail whenever R forces the number to be represented in scientific notation.
  There is some relief from this as \code{numberToWord} will attempt to resent the
  \code{scipen} option to prevent the use of scientific notation, but sometimes R insists on
  doing things its own way.  I'm sure there's a good, succinct explanation for when to expect
  this to fail, but since it is mostly likely to happen at preposterously large values (into the
  trillions), I'm not too concerned about because let's face it--at that point, you should
  probably reconsider if you really want to print out the number as text.  Do note, however,
  that exact powers of 10 that exceed one billion are problematic.
}
\examples{
numberToWord(1:10)
numberToWord(1549846132187489464)
}
\author{
Benjamin Nutter
}

