% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/percentileWtGestAge.R
\name{percentileWtGestAge}
\alias{percentileWtGestAge}
\title{Percentile Weight for Gestational Age}
\source{
Data are extracted from the World Health Organization (WHO) website (see link below).
  The link to download the MS Excel sheet is under the "Methodological aspects and other analysis"
  and titled "Weight percentiles calculator."

  \url{http://www.who.int/reproductivehealth/topics/maternal_perinatal/globalsurvey/en/}
}
\usage{
percentileWtGestAge(weeks, wt, less = TRUE, direction = FALSE,
  result = c("ptile", "code", "desc"))
}
\arguments{
\item{weeks}{Weeks gestational age (as an integer)}

\item{wt}{Weight in grams}

\item{less}{Logical.  Determines if the percentile returned should be read as
"less than [percentile]" or "greater than [percentile]"}

\item{direction}{Logical. Should an inequality symbol be printed with the
result to indicate the direction.  See Examples.}

\item{result}{A character specifying how the results should be returned.
  \code{"ptile"} results in the numeric percentile value being returned;
  \code{"code"} returns the percentile code; and
  \code{"desc"} returns the percentile description.  These values are
  represented by the \code{percentile}, \code{percentile_code}, and
  \code{percentile_desc} fields, respectively, in the \code{GestAgeWtPercentile}
  data set.}
}
\description{
Reference a birth weight for gestational age to the
  World Health Organization growth tables.
}
\examples{
#* Results can be read as "less than 75th percentile" (for example)
percentileWtGestAge(weeks = c(30, 34, 36),
                    wt = c(1500, 2500, 2000))

#* Same results, but with directional indicator
percentileWtGestAge(weeks = c(30, 34, 36),
                    wt = c(1500, 2500, 2000),
                    direction=TRUE)

#* Same results, but with directional indicator
#* and percentile description
percentileWtGestAge(weeks = c(30, 34, 36),
                    wt = c(1500, 2500, 2000),
                    direction=TRUE,
                    result="desc")
}
\author{
Benjamin Nutter
}

