% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapid_eda.R
\name{rapid_eda}
\alias{rapid_eda}
\title{Summarize Numeric Variables}
\usage{
rapid_eda(data, vars = NULL, group = NULL, quiet = FALSE)
}
\arguments{
\item{data}{A \code{data.frame}-like object.  Accepts objects of 
class \code{data.frame}, \code{tbl_df}, \code{grouped_df}, 
and \code{data.table}.}

\item{vars}{A character vector giving the names of columns to 
summarize.  When \code{NULL}, all numeric variables will be 
summarized.}

\item{group}{A character vector giving the names of columns by 
which the summaries are to be grouped.  When \code{NULL}, 
no grouping is performed.}

\item{quiet}{\code{logical(1)}. Should warnings be silenced.  See
Details.}
}
\value{
Returns an object of the same class as \code{data}.  Summarized
  values include:
\itemize{
  \item{n }{Total non-missing values}
  \item{n_miss }{Total missing values}
  \item{mean }{Arithmetic mean}
  \item{sd }{Standard deviation}
  \item{min }{Minimum}
  \item{p25 }{Twenty-fifth percentile (first quartile)}
  \item{median }{Median}
  \item{p75 }{Seventy-fifth percentile (third quartile)}
  \item{max }{Maximum}
}
}
\description{
Produce common numeric summary values for continuous 
  variables in a data frame.  The function only operates on 
  variables of class \code{numeric} and \code{integer}.
}
\details{
If \code{data} is either a \code{data.frame} or \code{data.table}, 
  it is first converted to a \code{tbl_df}.  It is converted back 
  to it's original class when returning the output.

  \code{rapid_eda} only summarizes data for numeric variables.
  To prevent failures, any non-numeric variables provided for summary
  will be dropped from \code{data} prior to generating the summary.
  A warning is printed naming the columns dropped unless 
  \code{quite = FALSE}
}
\section{Functional Requirements}{

\enumerate{
  \item Accepts the following classes as input: \code{data.frame}, 
        \code{tbl_df}, \code{grouped_df}, \code{data.table}
  \item Accepts a vector of column names to select for the summary.
  \item Accepts a vector of column names to group by
  \item When given a \code{groupd_df}, the grouping variables are 
        maintained.  Any new variables given to group by in the argument
        are added.
  \item Includes an option to silence warnings
  \item Returns a warning naming any non-numeric variables dropped
        from the summary.
  \item Returns a warning naming an variables requested but not found
        in \code{data}
  \item Returns an object that is the same class as the object 
        given to the function.
  \item Output lists variables in the same order as \code{vars}.  When
        \code{vars = NULL}, the order is assumed to be the same order
        they are given in the data.
}
}

\examples{
library(dplyr)

rapid_eda(mtcars)

rapid_eda(mtcars, 
          group = c("am", "gear"))

mtcars \%>\% 
  group_by(am, gear) \%>\%
  rapid_eda()

}
\author{
Benjamin Nutter
}
