% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_duplicated.R
\name{show_duplicated}
\alias{show_duplicated}
\title{Subset a Data Frame to Duplicated Subjects}
\usage{
show_duplicated(data, key)
}
\arguments{
\item{data}{A data frame}

\item{key}{A character vector naming the fields that identify unique
subjects (experimental units) in \code{data}.}
}
\description{
Reduces a data frame to those subjects that are duplicated,
  but also includes the index record.
}
\details{
Using the \code{duplicated} function returns a logical vector
  indicating which records are duplicates, but will show \code{FALSE}
  for the first record of the subject being duplicated.  
  \code{show_duplicated} returns all of the rows for the subject
  for inspection.
}
\section{Functional Requirements}{

\enumerate{
  \item Accept an object that inherits the \code{data.frame} class.
  \item Returns an object of the same class given.
  \item Gives all of the records for an experimental unit found to 
        have at least one duplication on \code{key} columns.
}
}

\examples{
DF <- data.frame(
  id = c(1, 1, 1, 2, 2, 2, 3, 3, 3),
  week = c(1, 1, 2, 1, 2, 2, 1, 2, 3),
  obs = rep(letters[1:3], 3)
)

show_duplicated(DF, c("id", "week"))

show_duplicated(DF, c("id"))

show_duplicated(DF, c("id", "week", "obs"))

}
\author{
Benjamin Nutter
}
