% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{tidy}
\alias{tidy}
\title{Tidy the result of a test into a summary data.frame}
\usage{
tidy(x, ...)
}
\arguments{
\item{x}{An object ot be converted into a tidy data.frame.}

\item{...}{extra arguments}
}
\description{
The output of tidy is always a tidy data.frame.  
  See \link[broom]{tidy} for details.
}
