% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.summary.survfit.R
\name{tidy.summary.survfit}
\alias{tidy.summary.survfit}
\title{Tidy a Survfit Summary}
\usage{
\method{tidy}{summary.survfit}(x, detail = TRUE, ...)
}
\arguments{
\item{x}{a \code{summary.survfit} object.}

\item{detail}{A logical value indicating if the detailed output should be tidied, 
or just the table of survival metrics (median survival, etc)}

\item{...}{Additional arguments to pass to other methods}
}
\description{
Tidier for \code{summary.survfit} objects.  This isn't greatly different
  from \code{tidy.survfit}, except that it allows for the use of the \code{times} 
  argument.  It also doesn't contain any information with regard to censoring.
}
\section{Functional Requirements}{

\enumerate{
  \item accept a \code{summary.survfit} object
  \item return a tidy \code{data.frame}
}
}
\author{
Benjamin Nutter
}

