% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_plural.R
\name{toggle_plural}
\alias{toggle_plural}
\title{Toggle Between Plural and Singular Word Choice}
\usage{
toggle_plural(x, singular, plural, ...)
}
\arguments{
\item{x}{\code{numeric(1)}}

\item{singular}{\code{character(1)}, defines the word choice to use if 
\code{x} is equal to 1.}

\item{plural}{\code{character(1)}, defines the word choice to use if
\code{x} is not equal to 1.}

\item{...}{Additional arguments to pass to \code{\link{format}}}
}
\value{
Returns a \code{character(1)} object.
}
\description{
When writing dynamic documents, it is not always known if a 
 value will be singular or plural.  Writing a document that accommodates
 both conditions can be challenging.  \code{toggle_plural} assists
 by giving the user a choice of words, and inserts the appropriate
 word based on the value being printed.
}
\section{Functional Requirements}{

\enumerate{
  \item Return an error if \code{x} is not numeric.
  \item Return an error if either \code{singular} or plural is not
    a character string.
  \item Return \code{singular} if \code{x} is equal to 1, else 
    return \code{plural}.
}
}

\examples{
toggle_plural(1, "was", "were")

toggle_plural(1, "is", "are")

}
\seealso{
\code{\link{toggle_text}}, \code{\link{format}}, \code{\link{all.equal}}, 
  \code{\link{isTRUE}}
}
\author{
Benjamin Nutter
}
