% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_text.R
\name{toggle_text}
\alias{toggle_text}
\title{Toggle Text Output Based on a Condition}
\usage{
toggle_text(condition, true, false)
}
\arguments{
\item{condition}{An expression that resolves to a \code{logical(1)}}

\item{true}{\code{character(1)}, the string to print when 
\code{condition} resolves to \code{TRUE}.}

\item{false}{\code{character(1)}, the string to print when 
\code{condition} resolves to \code{FALSE}.}
}
\value{
\code{character(1)} object.
}
\description{
Toggle between two text strings based on the evaluation 
  of a logical.  This is helpful for writing dynamic reports where 
  the interpretation may differ depending on the result of a 
  condition.
}
\section{Functional Requirements}{

\enumerate{
  \item Return an error if \code{condition} does not resolve to a logical.
  \item Return an error if either \code{true} or \code{false} are not 
    character strings.
  \item Return \code{true} when condition resolves to \code{TRUE}, else
    return \code{false}
}
}

\examples{
p_value <- 0.01
toggle_text(p_value <= 0.05, 
            true = "is statistically significant",
            false = "is not statistically significant")

}
\seealso{
\code{toggle_plural}
}
\author{
Benjamin Nutter
}
