% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariable_cat.R, R/univariable_cont.R
\name{univariable_cat}
\alias{univariable_cat}
\alias{univariable_cont}
\title{Univariable summary of a categorical variable.}
\usage{
univariable_cat(x, by = NULL, test = chisq.test, ..., ncat_warn = TRUE)

univariable_cont(x, by = NULL, test = wilcox.test, ..., ncat_warn = TRUE)
}
\description{

}
\section{Functional Requirements}{

\enumerate{
  \item Accept a vector of any type and treat it like categorical data
  \item Optionally accept a vector of any type the same length as 
    the first vector and treat it like categorical data
  \item Throw a warning when the grouping variable is not a character/factor
    or when there are more then 5 distinct categories.
  \item Accept a function name to be used for a univariable test.
  \item Accept additional arguments to the univariable test
}
}

