% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariable_table.R
\name{univariable_table}
\alias{univariable_table}
\title{Univariable Table Summary}
\usage{
univariable_table(data, vars, by = NULL, cat_vars = NULL, ...)
}
\arguments{
\item{data}{A data frame like object.}

\item{vars}{A character vector of variable names. The order listed here
determines the order the variables are listed in the table.}

\item{by}{A \code{character(1)} specifying a variable in \code{data} to 
use as a grouping variable.}

\item{...}{lists specifying tests and parameters to use for variables given in 
\code{vars}.  These are used to override the defaults of 
\code{univariable_cat} (\code{chisq.test}) and \code{univariable_cont} 
(\code{wilcox.test}).  See Specifying Tests.}
}
\description{
Performs a number of iterations to \code{univariable_cat} 
  and \code{univariable_cont} to construct a table of univariable summaries
  and comparisons. This is intended to streamline the generation of 
  "Table 1"s.
}
\section{Specifying Tests}{

There is no need to use \code{...} if you are satisfied with the default 
tests.  Each variable that requires a different test must have an entry in 
\code{...}.  Each entry must have an element named \code{test}. Additional
elements may be made to pass to the test.

If you would like to alter the arguments to a default test, you may 
pass the arguments without naming a \code{test} element.  

For a variable named \code{xvar} to be compared with a t-test, the 
entry in \code{...} would look like 
will look like: 
\code{xvar = list(test = t.test, alternative = 'greater')}
}

\section{Functional Requirements}{

\enumerate{
  \item Accept a data frame like object
  \item Accept a vector of variable name in \code{data} to summarize.
  \item Accept a \code{character(1)} giving a variable name to use as 
    a grouping variable.
  \item Accept a character vector of variables listed in \code{vars}
    to treat at categorical variables. (an override of numeric types)
  \item Gather all warnings and return them as one.
  \item Warnings should indicate the analysis in which the warning 
    was generated.
  \item Allow the user to designate alternate tests for each variable
    in \code{vars}
  \item Capture errors thrown during tests, return them as warnings
  \item Return NULL when an error occurs during a test
}
}
\examples{
univariable_table(data = mtcars,
                  vars = c("mpg", "vs", "qsec", "gear"),
                  by = "cyl",
                  cat_vars = c("vs", "gear"),
                  mpg = list(test = t.test, alternative = "greater"))

}

