% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_code.R
\name{vector_code}
\alias{vector_code}
\title{Print Vector to Console as R Code}
\usage{
vector_code(x, width = getOption("width"))
}
\arguments{
\item{x}{A vector}

\item{width}{The maximum width of each line of code.  By default, 
it uses the \code{width} option.}
}
\description{
Prints a vector to the console as R code that can be 
copied and pasted into a script.
}
\details{
This is intended as a code-writing utility.  As a convenience,
  the string returned to the console will have a width that will
  paste well into the user's code (instead of running off the 
  right side of the screen).  To maintain the width, efficiency has 
  been sacrificed.  It is assumed that this will not be used on 
  large vectors.
}
\section{Functional Requirements}{

\enumerate{
  \item Accept a vector of any type.
  \item Return a string suitable for a copy and paste into code.
  \item The string returned must provide the same class as the
        original vector (within limits: dates will be strings)
}
}
\examples{
vector_code(1:3)

}
\author{
Benjamin Nutter
}

