% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vectorCode.R
\name{vectorCode}
\alias{vectorCode}
\title{Print Vector to Console as R Code}
\usage{
vectorCode(x, width = getOption("width"))
}
\arguments{
\item{x}{A vector}

\item{width}{The maximum width of each line of code.  By default, it uses the
  \code{width} option.}
}
\description{
Prints a vector to the console as R code that can be copied and pasted
into a script.
}
\examples{
vectorCode(1:3)
}
\author{
Benjamin Nutter
}

