% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_warning.R
\name{with_warning}
\alias{with_warning}
\title{Evaluate an expression and capture any warning messages}
\usage{
with_warning(expr)
}
\arguments{
\item{expr}{An expression.}
}
\description{
Running a funtion like \code{univariable_table} has the 
  potential to produce several warnings, but can do so while making it 
  difficult to determine which of the variable tests generated the 
  warning.  This captures the warning without casting it so the warnings 
  can all be gathered and printed at the same time.
}

