% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_to_number.R
\name{word_to_number}
\alias{word_to_number}
\title{Convert Text-Represented Numbers to Their Numeric Representation}
\usage{
word_to_number(x)
}
\arguments{
\item{x}{A character vector. All punctuation and instances of " and "
will be stripped from the vector within the function. In other words,
"eighty seven", "eighty-seven", and "eighty and seven" will all 
produce the same result.}
}
\description{
Convert a text-based numeric descriptor into its numeric 
  form.  For instance, "three" becomes "3".  Capable of handling 
  the special cases like "twelve hundred".
}
\details{
The conversion method here recognizes that it is only necessary
  to translate anything between 0 and 999.  After that point, everything
  begins to repeat, and only an adjustment to the thousands, millions, 
  etc need be applied.  The translation, then, takes place in three steps.
  First, the number is broken into magnitude groups in which the hundreds
  may be translated. Second, each magnitude is translated. Finally, 
  the translation of each magnitude is multiplied by its magnitude's value
  and all of the results are summed.
  
  For example, the description "thirteen million three hundred and ninety 
  one thousand four hundred and twelve" is broken into three groups:
  
  \itemize{
    \item "thirteen million"
    \item "three hundred and ninety one thousand"
    \item "four hundred and twelve"
  }
  
  The translation of each becomes
  
  \itemize{
    \item 13 * 1000000
    \item 391 * 1000
    \item 412
  }
  
  Those three translations are added together to yield 13391412.
}
\section{Functional Requirements}{

\enumerate{
 \item Returns a numeric vector translate the contents of \code{x}
 \item Cast an error if \code{x} is not a character vector.
 \item Cast an error if any words in \code{x} do not match the 
   names of \code{c(word_to_number_reference, magnitude_reference)}
 \item Cleanses \code{x} of all punctuation and instances of the 
   string \code{" and "}.
 \item Cast a warning if any string failed to cast an error but 
   could not be translated.
}
}

\examples{
word_to_number(c("one hundred eighty six", 
                 "three hundred and seventy-one",
                 "twelve hundred eighty-eight"))
word_to_number("thirteen million three hundred and ninety one thousand four hundred and twelve")
word_to_number("eight hundred and eleven")
}
\seealso{
\code{\link{number_to_word}}
}
